move env stuff into its own file
